requirements.txt
aiohttp==3.6.2
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
discord.py==1.3.3
Flask==1.1.2
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
pylint==2.5.2
python-dotenv==0.13.0
redis==3.5.3
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
